Adds user functionality to the application
