+++ src/main/resources/application.properties
@@ -14,4 +14,3 @@ myapp.bearer-token=${SALLING_API_KEY}
 app.api-key=${OPENAI_API_KEY}
 app.token-expiration = 1800
 app.token-issuer=dat.kea
-#hejMathias2
